+++ src/gen_michelson.ml
@@ -916,6 +916,11 @@ let to_ir (model : M.model) : T.ir =
     | Mratdur  (c, t)         -> let b = T.Bratdur    in add_builtin b; T.Icall (get_fun_name b, [f c; f t], is_inline b)
 
 
+    (* utils *)
+
+    | Mdatefromtimestamp _ -> emit_error (UnsupportedTerm ("Mdatefromtimestamp"))
+
+
     (* quantifiers *)
 
     | Mforall _ -> emit_error (UnsupportedTerm ("Mforall"))