+++ src/gen_transform.ml
@@ -1854,6 +1854,7 @@ let replace_date_duration_by_timestamp (model : model) : model =
     | Mnow          -> mk (Unix.time () |> int_of_float |> Big_int.big_int_of_int)
     | Mplus  (a, b) -> mk (Big_int.add_big_int (g a) (g b))
     | Mminus (a, b) -> mk (Big_int.sub_big_int (g a) (g b))
+    | Mdatefromtimestamp v -> mk (g v)
     | Mtimestamp _  -> x
     | _ ->
       begin
@@ -1866,6 +1867,7 @@ let replace_date_duration_by_timestamp (model : model) : model =
       match mt.node, mt.type_ with
       | Mdate d,_      -> mk_mterm (Mtimestamp (Core.date_to_timestamp d)) ttimestamp
       | Mduration d, _ -> mk_mterm (Mint (Core.duration_to_timestamp d)) tint
+      | Mdatefromtimestamp _, _ -> to_timestamp mt
       | Mnow, _        -> mk_mterm (Mnow) ttimestamp
       | Mmult (a, b), t when is_duration t && is_rat a.type_ && is_duration b.type_ ->
         let a = aux a in