+++ src/gen_why3.ml
@@ -2338,6 +2338,8 @@ let rec map_mterm m ctx (mt : M.mterm) : loc_term =
     | Minttorat v -> Ttuple ([map_mterm m ctx v; loc_term (Tint (Big_int.big_int_of_int 1))])
     | Mratdur (r,t) -> Tapp (loc_term (Tvar "rat_dur"),[map_mterm m ctx r; map_mterm m ctx t])
 
+    (* others ---------------------------------------------------------- *)
+    | Mdatefromtimestamp _ -> assert false
 
     (* quantifiers ---------------------------------------------------------- *)